requirements.txt
aiohappyeyeballs==2.4.3
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
deepdiff==8.0.1
fastapi==0.115.5
frozenlist==1.5.0
h11==0.14.0
idna==3.10
multidict==6.1.0
orderly-set==5.2.2
propcache==0.2.0
psycopg2-binary==2.9.10
pydantic==2.10.1
pydantic_core==2.27.1
requests==2.32.3
setuptools==75.1.0
six==1.16.0
sniffio==1.3.1
sse-starlette==2.1.3
sseclient==0.0.27
starlette==0.41.3
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.1
wheel==0.44.0
yarl==1.18.0
